CORE-8643 Move AppTemplateWizardContentPanelAppearance and gss to themes package
